test scm controller gitlab adapter
